requirements.txt
requests>=2.28
pandas>=1.5
boto3>=1.26
snowflake-connector-python[pandas]>=3.0
pytest>=7.0
